Get rid of builder
