+++ src/gui/app.rs
@@ -1,30 +1,51 @@
-use super::{
-	builder::EFFECTS_LIST,
-	color_tiles::{ColorTiles, ColorTilesState},
-};
+use super::color_tiles::{ColorTiles, ColorTilesState};
+use super::options_tile::OptionsTile;
+use super::{color_tiles, effect_browser_tile, options_tile};
+use crate::gui::menu_bar;
+use crate::keyboard_manager;
 use crate::{
 	enums::{Effects, Message},
-	gui::dialog::alert,
+	gui::dialog::{alert, panic},
 };
+use fltk::enums::FrameType;
+use fltk::{app, enums::Font, group::Pack, prelude::*, window::Window};
 use fltk::{
 	browser::HoldBrowser,
 	dialog,
-	menu::Choice,
 	prelude::{BrowserExt, MenuExt},
 	text,
 };
 use serde::{Deserialize, Serialize};
 use serde_json::Result;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::mpsc;
+use std::sync::Arc;
+use std::time::Duration;
+use std::{panic, thread};
+use std::{path, str::FromStr, sync::mpsc::Sender};
 
-use std::{
-	path,
-	str::FromStr,
-	sync::{
-		atomic::{AtomicBool, Ordering},
-		mpsc::Sender,
-		Arc,
-	},
-};
+const WIDTH: i32 = 900;
+const HEIGHT: i32 = 480;
+
+pub fn screen_center() -> (i32, i32) {
+	((app::screen_size().0 / 2.0) as i32, (app::screen_size().1 / 2.0) as i32)
+}
+
+pub const EFFECTS_LIST: [&str; 13] = [
+	"Static",
+	"Breath",
+	"Smooth",
+	"LeftWave",
+	"RightWave",
+	"Lightning",
+	"AmbientLight",
+	"SmoothLeftWave",
+	"SmoothRightWave",
+	"LeftSwipe",
+	"RightSwipe",
+	"Disco",
+	"Christmas",
+];
 
 #[derive(Serialize, Deserialize)]
 struct Profile {
@@ -37,8 +58,7 @@ struct Profile {
 pub struct App {
 	pub color_tiles: ColorTiles,
 	pub effect_browser: HoldBrowser,
-	pub speed_choice: Choice,
-	pub brightness_choice: Choice,
+	pub options_tile: OptionsTile,
 	pub tx: Sender<Message>,
 	pub stop_signal: Arc<AtomicBool>,
 	pub buf: text::TextBuffer,
@@ -69,8 +89,8 @@ impl App {
 				if let Ok(profile) = parse_profile(&self.buf.text()) {
 					self.color_tiles.set_state(&profile.color_tiles_state, profile.effect);
 					self.effect_browser.select(EFFECTS_LIST.iter().position(|&val| val == profile.effect.to_string()).unwrap() as i32 + 1);
-					self.speed_choice.set_value(profile.speed);
-					self.brightness_choice.set_value(profile.brightness);
+					self.options_tile.speed_choice.set_value(profile.speed);
+					self.options_tile.brightness_choice.set_value(profile.brightness);
 					self.stop_signal.store(true, Ordering::SeqCst);
 					self.tx.send(Message::UpdateEffect { effect: profile.effect }).unwrap();
 				} else {
@@ -83,8 +103,8 @@ impl App {
 					self.tx.send(Message::Refresh).unwrap();
 				}
 			} else if !is_default {
-   					alert(800, 200, "File does not exist!");
-   				}
+				alert(800, 200, "File does not exist!");
+			}
 		} else {
 			self.stop_signal.store(true, Ordering::SeqCst);
 			self.tx.send(Message::Refresh).unwrap();
@@ -94,8 +114,8 @@ impl App {
 		let profile = Profile {
 			color_tiles_state: self.color_tiles.get_state(),
 			effect: Effects::from_str(EFFECTS_LIST[self.effect_browser.value() as usize - 1]).unwrap(),
-			speed: self.speed_choice.value(),
-			brightness: self.brightness_choice.value(),
+			speed: self.options_tile.speed_choice.value(),
+			brightness: self.options_tile.brightness_choice.value(),
 		};
 
 		self.buf.set_text(serde_json::to_string(&profile).unwrap().as_str());
@@ -113,5 +133,172 @@ impl App {
 
 		self.stop_signal.store(true, Ordering::SeqCst);
 		self.tx.send(Message::Refresh).unwrap();
+	}
+	pub fn start_ui(mut manager: keyboard_manager::KeyboardManager, tx: mpsc::Sender<Message>, stop_signal: Arc<AtomicBool>) -> fltk::window::Window {
+		panic::set_hook(Box::new(|info| {
+			if let Some(s) = info.payload().downcast_ref::<&str>() {
+				panic(800, 400, s);
+			} else {
+				panic(800, 400, &info.to_string());
+			}
+		}));
+
+		//UI
+		let mut win = Window::new(screen_center().0 - WIDTH / 2, screen_center().1 - HEIGHT / 2, WIDTH, HEIGHT, "Legion Keyboard RGB Control");
+		let mut color_picker_pack = Pack::new(0, 30, 540, 360, "");
+		let tiles = color_tiles::ColorTiles::new(&tx, stop_signal.clone());
+
+		color_picker_pack.add(&tiles.zones.left.exterior_tile);
+		color_picker_pack.add(&tiles.zones.center_left.exterior_tile);
+		color_picker_pack.add(&tiles.zones.center_right.exterior_tile);
+		color_picker_pack.add(&tiles.zones.right.exterior_tile);
+		color_picker_pack.add(&tiles.master.exterior_tile);
+		color_picker_pack.end();
+
+		let effect_browser_tile = effect_browser_tile::EffectBrowserTile::create(540, 30, &EFFECTS_LIST);
+		let effect_browser = effect_browser_tile.effect_browser;
+
+		let options_tile = options_tile::OptionsTile::create(540, 390, &tx, &stop_signal.clone());
+
+		let mut app = Self {
+			color_tiles: tiles,
+			effect_browser,
+			options_tile,
+			tx,
+			stop_signal,
+			buf: text::TextBuffer::default(),
+			center: screen_center(),
+		};
+
+		menu_bar::AppMenuBar::new(&app.tx, app.stop_signal.clone(), &app);
+
+		win.end();
+		win.make_resizable(false);
+		win.show();
+
+		// Theming
+		app::background(51, 51, 51);
+		app::set_visible_focus(false);
+		app::set_font(Font::HelveticaBold);
+		app::set_frame_type(FrameType::FlatBox);
+
+		// Effect choice
+		app.effect_browser.set_callback({
+			let stop_signal = app.stop_signal.clone();
+			let tx = app.tx.clone();
+			let mut color_tiles = app.color_tiles.clone();
+			move |browser| {
+				stop_signal.store(true, Ordering::SeqCst);
+				match browser.value() {
+					0 => {
+						browser.select(0);
+					}
+					1 => {
+						color_tiles.update(Effects::Static);
+						tx.send(Message::UpdateEffect { effect: Effects::Static }).unwrap();
+					}
+					2 => {
+						color_tiles.update(Effects::Breath);
+						tx.send(Message::UpdateEffect { effect: Effects::Breath }).unwrap();
+					}
+					3 => {
+						color_tiles.update(Effects::Smooth);
+						tx.send(Message::UpdateEffect { effect: Effects::Smooth }).unwrap();
+					}
+					4 => {
+						color_tiles.update(Effects::LeftWave);
+						tx.send(Message::UpdateEffect { effect: Effects::LeftWave }).unwrap();
+					}
+					5 => {
+						color_tiles.update(Effects::RightWave);
+						tx.send(Message::UpdateEffect { effect: Effects::RightWave }).unwrap();
+					}
+					6 => {
+						color_tiles.update(Effects::Lightning);
+						tx.send(Message::UpdateEffect { effect: Effects::Lightning }).unwrap();
+					}
+					7 => {
+						color_tiles.update(Effects::AmbientLight);
+						tx.send(Message::UpdateEffect { effect: Effects::AmbientLight }).unwrap();
+					}
+					8 => {
+						color_tiles.update(Effects::SmoothLeftWave);
+						tx.send(Message::UpdateEffect { effect: Effects::SmoothLeftWave }).unwrap();
+					}
+					9 => {
+						color_tiles.update(Effects::SmoothRightWave);
+						tx.send(Message::UpdateEffect { effect: Effects::SmoothRightWave }).unwrap();
+					}
+					10 => {
+						color_tiles.update(Effects::LeftSwipe);
+						tx.send(Message::UpdateEffect { effect: Effects::LeftSwipe }).unwrap();
+					}
+					11 => {
+						color_tiles.update(Effects::RightSwipe);
+						tx.send(Message::UpdateEffect { effect: Effects::RightSwipe }).unwrap();
+					}
+					12 => {
+						color_tiles.update(Effects::Disco);
+						tx.send(Message::UpdateEffect { effect: Effects::Disco }).unwrap();
+					}
+					13 => {
+						color_tiles.update(Effects::Christmas);
+						tx.send(Message::UpdateEffect { effect: Effects::Christmas }).unwrap();
+					}
+					_ => {}
+				}
+			}
+		});
+
+		app.load_profile(true);
+
+		thread::spawn(move || loop {
+			match manager.rx.try_iter().last() {
+				Some(message) => {
+					match message {
+						Message::UpdateEffect { effect } => {
+							let color_array = app.color_tiles.get_zone_values();
+							let speed = app.options_tile.speed_choice.choice().unwrap().parse::<u8>().unwrap();
+							let brightness = app.options_tile.brightness_choice.choice().unwrap().parse::<u8>().unwrap();
+							manager.set_effect(effect, &color_array, speed, brightness);
+						}
+						Message::UpdateAllValues { value } => {
+							manager.keyboard.set_colors_to(&value);
+						}
+						Message::UpdateRGB { index, value } => {
+							manager.keyboard.solid_set_value_by_index(index, value);
+						}
+						Message::UpdateZone { zone_index, value } => {
+							manager.keyboard.set_zone_by_index(zone_index, value);
+						}
+						Message::UpdateValue { index, value } => {
+							manager.keyboard.set_value_by_index(index, value);
+						}
+						Message::UpdateBrightness { brightness } => {
+							manager.keyboard.set_brightness(brightness);
+							app.tx.send(Message::Refresh).unwrap();
+						}
+						Message::UpdateSpeed { speed } => {
+							manager.keyboard.set_speed(speed);
+							app.tx.send(Message::Refresh).unwrap();
+						}
+						Message::Refresh => {
+							app.tx.send(Message::UpdateEffect { effect: manager.last_effect }).unwrap();
+						}
+						Message::SaveProfile => {
+							app.save_profile();
+						}
+						Message::LoadProfile => {
+							app.load_profile(false);
+						}
+					}
+					app::awake();
+				}
+				None => {
+					thread::sleep(Duration::from_millis(20));
+				}
+			}
+		});
+		win
 	}
 }
